refactor(RoomescapeRepository): 중복 예약 검증 로직을 서비스 레이어로 이동
