+++ src/main/java/roomescape/service/RoomescapeService.java
@@ -19,10 +19,22 @@ public class RoomescapeService {
     }
 
     public Reservation addReservation(final Reservation reservation) {
+        if (existsSameReservation(reservation)) {
+            throw new IllegalArgumentException("[ERROR] 이미 존재하는 예약시간입니다.");
+        }
         return roomescapeRepository.saveReservation(reservation);
     }
 
     public void removeReservation(final long id) {
         roomescapeRepository.deleteById(id);
+    }
+
+    private boolean existsSameReservation(final Reservation reservation) {
+        List<Reservation> reservations = findReservations();
+        boolean exists = false;
+        for (Reservation candidate : reservations) {
+            exists = candidate.isDuplicateReservation(reservation);
+        }
+        return exists;
     }
 }
